refactor(rust): feature-gate heavy deps so core builds with --no-default-features
Make arrow/parquet/rayon/dashmap optional behind policy_db/parallel features,
move rand_distr to always-on, gate the modules that use them, and split the ort
OnnxEvaluator out of the always-compiled evaluator core. python/binary builds
unchanged; quoridor-rs core now builds with --no-default-features for wasm.
