+++ deep_quoridor/rust/src/agents/alphazero/mod.rs
@@ -3,12 +3,19 @@
 //! This module provides MCTS-based action selection using neural network evaluation.
 //! Only available behind the `binary` feature flag.
 
-pub mod eval_pipeline;
 pub mod evaluator;
 pub mod mcts;
-pub mod selfplay_game;
+
+#[cfg(feature = "binary")]
+pub mod agent;
+#[cfg(feature = "binary")]
+pub mod eval_pipeline;
+#[cfg(feature = "binary")]
 pub mod selfplay_mcts;
+#[cfg(feature = "binary")]
+pub mod selfplay_game;
+#[cfg(feature = "binary")]
 pub mod selfplay_metrics;
 
-pub mod agent;
+#[cfg(feature = "binary")]
 pub use agent::{AlphaZeroAgent, AlphaZeroAgentConfig};
