+++ deep_quoridor/rust/src/compact/mod.rs
@@ -1,6 +1,8 @@
 /// The algorithms in this directory make use of the QBitRepr compact state representation
+#[cfg(feature = "policy_db")]
 pub mod policy_db;
 pub mod q_bit_repr;
 mod q_bit_repr_conversions;
 pub mod q_game_mechanics;
+#[cfg(feature = "parallel")]
 pub mod q_minimax;
